+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Update NYSE and New York FED calendars following [Juneteenth National Independence Day Act](https://www.cnbc.com/2021/06/17/juneteenth-federal-holiday-biden-signs-bill.html)
+  (#98).
+
 ### Fixed
 
 ### Deprecated
@@ -16,6 +19,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Internal
 
 - Upgrade to parent [2.6.1](https://github.com/marcwrobel/parent/releases/tag/v2.6.1) (#94).
-- Upgrade to strata [2.12.3](https://strata.opengamma.io/releases/) (#96).
+- Upgrade to strata [2.12.4](https://strata.opengamma.io/releases/) (#96, #99).
 
 ### Thanks